Profile page + Tests + Nice pages + Form validators (#21)
* token csrf in login

* profile page is working

* voter for delete added

* validators added

* validators added

* comments ajax route added

* tests for the entities

* config changed due to bugs in webtestcase

* lot of tests today

* tests and frontend

* nice data fixtures

* data fixture update

* MPD

* navbar responsive

* code reviewed

* see medias responsive

+++ src/Controller/TrickController.php
@@ -16,11 +16,13 @@ use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 class TrickController extends AbstractController
 {
     /**
      * @Route("/trick/new", name="trick_new", methods={"GET","POST"})
+     * @IsGranted("ROLE_USER")
      */
     public function new(Request $request, FileUploader $fileUploader, UserRepository $userRepository): Response
     {
@@ -30,13 +32,13 @@ class TrickController extends AbstractController
         $form = $this->createForm(TrickType::class, $trick);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->persistValidForm($form, $fileUploader, $trick, $author);
+            $this->persistValidForm($request, $form, $fileUploader, $trick, $author);
             return $this->redirectToRoute('home');
         }
 
         return $this->render('trick/new.html.twig', [
             'trick' => $trick,
-            'form' => $form->createView(),
+            'form' => $form->createView()
         ]);
     }
 
@@ -46,7 +48,7 @@ class TrickController extends AbstractController
     public function show(Trick $trick, Request $request, CommentRepository $commentRepository): Response
     {
         $em = $this->getDoctrine()->getManager();
-        $comments = $commentRepository->findBy(['trick'=>$trick]);
+        $comments = $commentRepository->findBy(['trick' => $trick], ["createdAt" => "DESC"]);
 
         $newComment = new Comment();
         $formComment = $this->createForm(CommentType::class, $newComment);
@@ -57,23 +59,25 @@ class TrickController extends AbstractController
             $newComment->setTrick($trick);
             $em->persist($newComment);
             $em->flush();
-            return $this->redirectToRoute('trick_show',['id'=>$trick->getId(),'slug'=>$trick->getSlug()]);
+            $this->addFlash('notice', 'Commentaire ajouté');
+            return $this->redirectToRoute('trick_show', ['id' => $trick->getId(), 'slug' => $trick->getSlug()]);
         }
 
         return $this->render('trick/show.html.twig', [
             'trick' => $trick,
             'formComment' => $formComment->createView(),
-            'comments' => $comments 
+            'comments' => $comments
         ]);
     }
 
     /**
      * @Route("/trick/{id}/edit", name="trick_edit", methods={"GET","POST"}, requirements={"id":"\d+"})
+     * @IsGranted("ROLE_USER")
      */
     public function edit(Request $request, Trick $trick, FileUploader $fileUploader, ImageRepository $imageRepository): Response
     {
-        $em = $this->getDoctrine()->getManager();
         $images = $imageRepository->findBy(['trick' => $trick]);
+
         $arrayPhotoNames = null;
         foreach ($images as $image) {
             $arrayPhotoNames[] = $image->getPath();
@@ -84,7 +88,7 @@ class TrickController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->persistValidForm($form, $fileUploader, $trick, null, $arrayPhotoNames);
+            $this->persistValidForm($request, $form, $fileUploader, $trick, null, $arrayPhotoNames);
             return $this->redirectToRoute('home');
         }
 
@@ -96,21 +100,31 @@ class TrickController extends AbstractController
     }
 
     /**
-     * @Route("/trick/{id}", name="trick_delete", methods={"POST"})
+     * @Route("/trick/{id}/delete", name="trick_delete", requirements={"id":"\d+"})
      */
     public function delete(Request $request, Trick $trick, ImageRepository $imageRepository): Response
     {
+        // check delete voter
+        $this->denyAccessUnlessGranted('trick_delete', $trick);
+
         if ($this->isCsrfTokenValid('delete' . $trick->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             // Delete all images linked to the trick
             $images = $imageRepository->findBy(['trick' => $trick]);
             foreach ($images as $image) {
-                $fileToDelete = new File($this->getParameter('trickUpload_directory') . "/" . $image->getPath());
-                $this->deleteFile($fileToDelete);
+                // the images in the ina_array are the images that we should'nt delete, they are used in the fixture
+                if ($image->getPath() && !in_array($image->getPath(),['snowboard_main.jpeg','180.jpeg','360.jpeg', '540.jpeg', '1080.jpeg', 'tailSlide.jpeg', 'japan.jpeg', 'nosegrab.jpeg', 'mactwist.jpeg', 'mute.jpeg', 'sad.jpeg', 'indy.jpeg'])) {
+                    if (file_exists($this->getParameter('trickUpload_directory') . "/" . $image->getPath())) {
+                        $fileToDelete = new File($this->getParameter('trickUpload_directory') . "/" . $image->getPath());
+                        $this->deleteFile($fileToDelete);
+                    }
+                }
             }
-            if ($trick->getMainImgName()) {
-                $mainImgTodelete = new File($this->getParameter('trickUpload_directory') . "/" . $trick->getMainImgName());
-                $this->deleteFile($mainImgTodelete);
+            if ($trick->getMainImgName() && $trick->getMainImgName() !== "snowboard_main.jpeg") {
+                if (file_exists($this->getParameter('trickUpload_directory') . "/" . $trick->getMainImgName())) {
+                    $mainImgTodelete = new File($this->getParameter('trickUpload_directory') . "/" . $trick->getMainImgName());
+                    $this->deleteFile($mainImgTodelete);
+                }
             }
 
             $em->remove($trick);
@@ -120,22 +134,26 @@ class TrickController extends AbstractController
         return $this->redirectToRoute('home');
     }
 
-    private function persistValidForm($form, $fileUploader, $trick, $author = null, $oldImages = null)
+    private function persistValidForm($request, $form, $fileUploader, $trick, $author = null, $oldImages = null)
     {
         $em = $this->getDoctrine()->getManager();
         /** @var UploadedFile $mainImgFile */
         $mainImgFile = $form->get('mainImg')->getData();
 
+        $mainImgSrcData = $request->request->get('mainImgSrcData');
+
         $videos = $form->get('videos')->getData();
 
         if ($mainImgFile) {
             $oldFile = null;
-            if ($trick->getMainImgName()) {
-                $oldFile = new File($this->getParameter('trickUpload_directory') . "/" . $trick->getMainImgName());
+            if ($trick->getMainImgName() && $trick->getMainImgName() !== "snowboard_main.jpeg") {
+                if (file_exists($this->getParameter('trickUpload_directory') . "/" . $trick->getMainImgName())) {
+                    $oldFile = new File($this->getParameter('trickUpload_directory') . "/" . $trick->getMainImgName());
+                }
             }
             $newFilename = $fileUploader->upload($mainImgFile, $this->getParameter('trickUpload_directory'), $oldFile);
             $trick->setMainImgName($newFilename);
-        }else if (!$trick->getMainImgName()){
+        } else if (!$trick->getMainImgName() || $mainImgSrcData === "1") {
             $trick->setMainImgName('snowboard_main.jpeg');
         }
 
@@ -144,16 +162,18 @@ class TrickController extends AbstractController
         $newImages = [];
         if ($images) {
             foreach ($images as $image) {
-                if ($image->getFile() !== null) {
-                    $newFilename = $fileUploader->upload($image->getFile(), $this->getParameter('trickUpload_directory'));
-                    $newImages[] = $newFilename;
+                if ($image->getFile() || $image->getPath()) {
+                    if ($image->getFile() !== null) {
+                        $newFilename = $fileUploader->upload($image->getFile(), $this->getParameter('trickUpload_directory'));
+                        $newImages[] = $newFilename;
+                    }
+                    if ($image->getFile() === null) {
+                        $newFilename = $image->getPath();
+                        $newImages[] = $newFilename;
+                    }
+                    $image->setPath($newFilename);
+                    $image->setTrick($trick);
                 }
-                if ($image->getFile() === null) {
-                    $newFilename = $image->getPath();
-                    $newImages[] = $newFilename;
-                }
-                $image->setPath($newFilename);
-                $image->setTrick($trick);
             }
         }
 
@@ -161,8 +181,10 @@ class TrickController extends AbstractController
         if ($oldImages) {
             $imagesToDelete = array_diff($oldImages, $newImages);
             foreach ($imagesToDelete as $imageToDelete) {
-                $fileToDelete = new File($this->getParameter('trickUpload_directory') . "/" . $imageToDelete);
-                $this->deleteFile($fileToDelete);
+                if (file_exists($this->getParameter('trickUpload_directory') . "/" . $imageToDelete)) {
+                    $fileToDelete = new File($this->getParameter('trickUpload_directory') . "/" . $imageToDelete);
+                    $this->deleteFile($fileToDelete);
+                }
             }
         }
 
@@ -182,5 +204,23 @@ class TrickController extends AbstractController
         if (file_exists($fileToDelete->getPathname())) {
             unlink($fileToDelete);
         }
+    }
+
+    /**
+     * @Route("/getMoreComments", name="getMoreComments", methods={"POST"})
+     */
+    public function getMoreComments(Request $request, CommentRepository $commentRepository)
+    {
+        $offset = $request->request->get('offset');
+        $newComments = $commentRepository->findBy([], null, 5, $offset);
+
+        return $this->json(
+            [
+                'newComments' => $newComments
+            ],
+            200,
+            [],
+            ['groups' => 'comment']
+        );
     }
 }
